requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_27zk0eqdh0/croot/brotli-split_1714483157007/work
certifi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_0ecs_ou28u/croot/certifi_1725551740278/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_06ndtibm2c/croot/cffi_1726856446111/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1728479282467/work
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
h5py==3.12.1
hpack==4.0.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1729190692267/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1717461632069/work
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.24.3
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
proglog @ file:///home/conda/feedstock_root/build_artifacts/proglog_1728452852274/work
protobuf==4.25.5
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
Pygments==2.18.0
pyparsing==3.2.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
rich==13.9.4
six==1.16.0
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==2.5.0
torch==2.2.2
torchvision==0.17.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1732497199771/work
typing_extensions==4.12.2
tzdata==2024.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1726496430923/work
Werkzeug==3.1.3
wrapt==1.17.0
zstandard @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a5_i6g4o6n/croot/zstandard_1731356352787/work
